+++ server/src/main/java/org/candlepin/hostedtest/AdapterOverrideModule.java
@@ -14,6 +14,7 @@
  */
 package org.candlepin.hostedtest;
 
+import org.candlepin.service.ProductServiceAdapter;
 import org.candlepin.service.SubscriptionServiceAdapter;
 
 import com.google.inject.AbstractModule;
@@ -32,10 +33,13 @@ public class AdapterOverrideModule extends AbstractModule {
      */
     @Override
     protected void configure() {
+        bind(HostedTestResource.class);
+        bind(HostedTestDataStore.class).asEagerSingleton();
 
         bind(SubscriptionServiceAdapter.class).to(HostedTestSubscriptionServiceAdapter.class)
-                .asEagerSingleton();
-        bind(HostedTestSubscriptionResource.class);
+            .asEagerSingleton();
+        bind(ProductServiceAdapter.class).to(HostedTestProductServiceAdapter.class)
+            .asEagerSingleton();
     }
 
 }
