+++ server/src/main/java/org/candlepin/hostedtest/HostedTestResource.java
@@ -54,17 +54,15 @@ import javax.ws.rs.core.MediaType;
 
 
 /**
- * The SubscriptionResource class is used to provide an
- * in-memory upstream source for subscriptions when candlepin is run in hosted
- * mode, while it is built with candlepin, it is not packaged in candlepin.war,
- * as the only purpose of this class is to support spec tests.
+ * The HostedTestResource class provides an endpoint for managing the upstream data stored by the
+ * backing HostedTestDataStore class
  */
 @SuppressSwaggerCheck
 @Path("/hostedtest")
-public class HostedTestSubscriptionResource {
+public class HostedTestResource {
 
     @Inject
-    private HostedTestSubscriptionServiceAdapter adapter;
+    private HostedTestDataStore datastore;
 
     @Inject
     private UniqueIdGenerator idGenerator;
@@ -119,21 +117,21 @@ public class HostedTestSubscriptionResource {
 
         // Create content...
         for (ContentData content : cmap.values()) {
-            if (this.adapter.getContent(content.getId()) != null) {
-                this.adapter.updateContent(content.getId(), content);
+            if (this.datastore.getContent(content.getId()) != null) {
+                this.datastore.updateContent(content.getId(), content);
             }
             else {
-                this.adapter.createContent(content);
+                this.datastore.createContent(content);
             }
         }
 
         // Create products...
         for (ProductData product : pmap.values()) {
-            if (this.adapter.getProduct(product.getId()) != null) {
-                this.adapter.updateProduct(product.getId(), product);
+            if (this.datastore.getProduct(product.getId()) != null) {
+                this.datastore.updateProduct(product.getId(), product);
             }
             else {
-                this.adapter.createProduct(product);
+                this.datastore.createProduct(product);
             }
         }
     }
@@ -201,7 +199,7 @@ public class HostedTestSubscriptionResource {
             subscription.setId(this.idGenerator.generateId());
         }
 
-        if (this.adapter.getSubscription(subscription.getId()) != null) {
+        if (this.datastore.getSubscription(subscription.getId()) != null) {
             throw new ConflictException("subscription already exists: " + subscription.getId());
         }
 
@@ -209,7 +207,7 @@ public class HostedTestSubscriptionResource {
         this.createSubscriptionObjects(subscription);
 
         // Create subscription object...
-        SubscriptionInfo sinfo = this.adapter.createSubscription(subscription);
+        SubscriptionInfo sinfo = this.datastore.createSubscription(subscription);
 
         return sinfo;
     }
@@ -224,7 +222,7 @@ public class HostedTestSubscriptionResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/subscriptions")
     public Collection<? extends SubscriptionInfo> listSubscriptions() {
-        return this.adapter.getSubscriptions();
+        return this.datastore.listSubscriptions();
     }
 
     /**
@@ -241,7 +239,7 @@ public class HostedTestSubscriptionResource {
     @Path("/subscriptions/{subscription_id}")
     @Produces(MediaType.APPLICATION_JSON)
     public SubscriptionInfo getSubscription(@PathParam("subscription_id") String subscriptionId) {
-        return adapter.getSubscription(subscriptionId);
+        return this.datastore.getSubscription(subscriptionId);
     }
 
     /**
@@ -267,7 +265,7 @@ public class HostedTestSubscriptionResource {
             throw new BadRequestException("no subscription data provided");
         }
 
-        if (this.adapter.getSubscription(subscriptionId) == null) {
+        if (this.datastore.getSubscription(subscriptionId) == null) {
             throw new NotFoundException("subscription does not yet exist: " + subscriptionId);
         }
 
@@ -275,7 +273,7 @@ public class HostedTestSubscriptionResource {
         this.createSubscriptionObjects(subscription);
 
         // Update subscription
-        return this.adapter.updateSubscription(subscriptionId, subscription);
+        return this.datastore.updateSubscription(subscriptionId, subscription);
     }
 
     /**
@@ -291,7 +289,7 @@ public class HostedTestSubscriptionResource {
     @Path("/subscriptions/{subscription_id}")
     @Produces(MediaType.APPLICATION_JSON)
     public boolean deleteSubscription(@PathParam("subscription_id") String subscriptionId) {
-        return adapter.deleteSubscription(subscriptionId) != null;
+        return this.datastore.deleteSubscription(subscriptionId) != null;
     }
 
     /**
@@ -300,7 +298,7 @@ public class HostedTestSubscriptionResource {
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
     public void clearData() {
-        this.adapter.clearData();
+        this.datastore.clearData();
     }
 
     @POST
@@ -312,14 +310,14 @@ public class HostedTestSubscriptionResource {
         @PathParam("product_id") String productId,
         Map<String, Boolean> contentMap) {
 
-        ProductInfo pinfo = this.adapter.getProduct(productId);
+        ProductInfo pinfo = this.datastore.getProduct(productId);
 
         if (pinfo == null) {
             throw new NotFoundException("product not found: " + productId);
         }
 
         for (String contentId : contentMap.keySet()) {
-            if (this.adapter.getContent(contentId) == null) {
+            if (this.datastore.getContent(contentId) == null) {
                 throw new NotFoundException("content not found: " + contentId);
             }
         }
@@ -331,7 +329,7 @@ public class HostedTestSubscriptionResource {
                 entry.getValue() :
                 ProductContent.DEFAULT_ENABLED_STATE;
 
-            this.adapter.addContentToProduct(productId, contentId, enabled);
+            this.datastore.addContentToProduct(productId, contentId, enabled);
         }
 
         return pinfo;
@@ -359,17 +357,17 @@ public class HostedTestSubscriptionResource {
     public ProductInfo removeContentFromProduct(@PathParam("product_id") String productId,
         Collection<String> contentIds) {
 
-        if (this.adapter.getProduct(productId) == null) {
+        if (this.datastore.getProduct(productId) == null) {
             throw new NotFoundException("product not found: " + productId);
         }
 
         if (contentIds != null) {
             for (String contentId : contentIds) {
-                this.adapter.removeContentFromProduct(productId, contentId);
+                this.datastore.removeContentFromProduct(productId, contentId);
             }
         }
 
-        return this.adapter.getProduct(productId);
+        return this.datastore.getProduct(productId);
     }
 
     @DELETE
@@ -386,7 +384,7 @@ public class HostedTestSubscriptionResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Transactional
     public Collection<? extends ProductInfo> listProducts() {
-        return this.adapter.listProducts();
+        return this.datastore.listProducts();
     }
 
     @GET
@@ -394,7 +392,7 @@ public class HostedTestSubscriptionResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Transactional
     public ProductInfo getProduct(@PathParam("product_id") String productId) {
-        ProductInfo pinfo = this.adapter.getProduct(productId);
+        ProductInfo pinfo = this.datastore.getProduct(productId);
 
         if (pinfo == null) {
             throw new NotFoundException("product does not exist: " + productId);
@@ -417,11 +415,11 @@ public class HostedTestSubscriptionResource {
             throw new BadRequestException("product lacks a product ID: " + product);
         }
 
-        if (this.adapter.getProduct(product.getId()) != null) {
+        if (this.datastore.getProduct(product.getId()) != null) {
             throw new ConflictException("product already exists: " + product.getId());
         }
 
-        return this.adapter.createProduct(product);
+        return this.datastore.createProduct(product);
     }
 
     @PUT
@@ -437,18 +435,18 @@ public class HostedTestSubscriptionResource {
             throw new BadRequestException("product update is null");
         }
 
-        if (this.adapter.getProduct(productId) == null) {
+        if (this.datastore.getProduct(productId) == null) {
             throw new NotFoundException("product does not yet exist: " + productId);
         }
 
-        return this.adapter.updateProduct(productId, update);
+        return this.datastore.updateProduct(productId, update);
     }
 
     @DELETE
     @Path("/products/{product_id}")
     @Produces(MediaType.APPLICATION_JSON)
     public boolean deleteProduct(@PathParam("product_id") String productId) {
-        return adapter.deleteProduct(productId) != null;
+        return this.datastore.deleteProduct(productId) != null;
     }
 
 
@@ -457,7 +455,7 @@ public class HostedTestSubscriptionResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Transactional
     public Collection<? extends ContentInfo> listContent() {
-        return this.adapter.listContent();
+        return this.datastore.listContent();
     }
 
     @GET
@@ -465,7 +463,7 @@ public class HostedTestSubscriptionResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Transactional
     public ContentInfo getContent(@PathParam("content_id") String contentId) {
-        ContentInfo cinfo = this.adapter.getContent(contentId);
+        ContentInfo cinfo = this.datastore.getContent(contentId);
 
         if (cinfo == null) {
             throw new NotFoundException("content does not exist: " + contentId);
@@ -488,11 +486,11 @@ public class HostedTestSubscriptionResource {
             throw new BadRequestException("content lacks a content ID: " + content);
         }
 
-        if (this.adapter.getContent(content.getId()) != null) {
+        if (this.datastore.getContent(content.getId()) != null) {
             throw new ConflictException("content already exists: " + content.getId());
         }
 
-        return this.adapter.createContent(content);
+        return this.datastore.createContent(content);
     }
 
     @PUT
@@ -508,18 +506,18 @@ public class HostedTestSubscriptionResource {
             throw new BadRequestException("content update is null");
         }
 
-        if (this.adapter.getContent(contentId) == null) {
+        if (this.datastore.getContent(contentId) == null) {
             throw new NotFoundException("content does not yet exist: " + contentId);
         }
 
-        return this.adapter.updateContent(contentId, update);
+        return this.datastore.updateContent(contentId, update);
     }
 
     @DELETE
     @Path("/content/{content_id}")
     @Produces(MediaType.APPLICATION_JSON)
     public boolean deleteContent(@PathParam("content_id") String contentId) {
-        return adapter.deleteContent(contentId) != null;
+        return this.datastore.deleteContent(contentId) != null;
     }
 
 }
