+++ server/src/test/java/org/candlepin/controller/PoolManagerTest.java
@@ -89,6 +89,7 @@ import org.candlepin.policy.js.entitlement.PreUnbindHelper;
 import org.candlepin.policy.js.pool.PoolRules;
 import org.candlepin.policy.js.pool.PoolUpdate;
 import org.candlepin.resource.dto.AutobindData;
+import org.candlepin.service.ProductServiceAdapter;
 import org.candlepin.service.SubscriptionServiceAdapter;
 import org.candlepin.service.model.SubscriptionInfo;
 import org.candlepin.test.MockResultIterator;
@@ -139,6 +140,7 @@ public class PoolManagerTest {
     private I18n i18n;
 
     @Mock private PoolCurator mockPoolCurator;
+    @Mock private ProductServiceAdapter mockProdAdapter;
     @Mock private SubscriptionServiceAdapter mockSubAdapter;
     @Mock private ProductCurator mockProductCurator;
     @Mock private ProductManager mockProductManager;
@@ -434,7 +436,7 @@ public class PoolManagerTest {
 
         when(mockPoolCurator.listByOwnerAndTypes(eq(owner.getId()), any(PoolType.class))).thenReturn(pools);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         List<Pool> expectedFloating = new LinkedList();
 
         // Make sure that only the floating pool was regenerated
@@ -472,7 +474,7 @@ public class PoolManagerTest {
 
         when(mockPoolCurator.listByOwnerAndTypes(eq(owner.getId()), any(PoolType.class))).thenReturn(pools);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         List<Pool> expectedModified = new LinkedList();
 
         // Make sure that only the floating pool was regenerated
@@ -747,7 +749,7 @@ public class PoolManagerTest {
 
         Owner owner = getOwner();
         when(mockOwnerCurator.getByKey(owner.getKey())).thenReturn(owner);
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         List<Pool> poolsToDelete = Arrays.asList(p);
         verify(this.manager).deletePools(eq(poolsToDelete));
     }
@@ -780,7 +782,7 @@ public class PoolManagerTest {
         when(cqmock.list()).thenReturn(Collections.<Pool>emptyList());
         when(mockPoolCurator.getPoolsBySubscriptionIds(anyList())).thenReturn(cqmock);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         List<Pool> delPools = Arrays.asList(p);
         verify(this.manager).deletePools(eq(delPools));
     }
@@ -809,7 +811,7 @@ public class PoolManagerTest {
 
         when(mockPoolCurator.listByOwnerAndTypes(eq(owner.getId()), any(PoolType.class))).thenReturn(pools);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         verify(this.manager, never()).deletePool(same(p));
     }
 
@@ -836,7 +838,7 @@ public class PoolManagerTest {
 
         Owner owner = getOwner();
         when(mockOwnerCurator.getByKey(owner.getKey())).thenReturn(owner);
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         verify(this.manager, never()).deletePool(same(p));
     }
 
@@ -862,7 +864,7 @@ public class PoolManagerTest {
         Owner owner = getOwner();
         when(mockOwnerCurator.getByKey(owner.getKey())).thenReturn(owner);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         verify(this.manager, never()).deletePool(same(p));
     }
 
@@ -888,7 +890,7 @@ public class PoolManagerTest {
 
         when(mockPoolCurator.listByOwnerAndTypes(eq(owner.getId()), any(PoolType.class))).thenReturn(pools);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         ArgumentCaptor<List> poolCaptor = ArgumentCaptor.forClass(List.class);
         verify(this.poolRulesMock).updatePools(poolCaptor.capture(), any(Map.class));
         assertEquals(1, poolCaptor.getValue().size());
@@ -931,7 +933,7 @@ public class PoolManagerTest {
         when(cqmock.list()).thenReturn(Collections.<Pool>emptyList());
         when(mockPoolCurator.getPoolsBySubscriptionId(anyString())).thenReturn(cqmock);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
 
         assertPoolsAreEqual(TestUtil.copyFromSub(s), argPool.getValue());
         verify(this.mockPoolCurator, times(1)).create(any(Pool.class));
@@ -975,7 +977,7 @@ public class PoolManagerTest {
 
         when(mockPoolCurator.listByOwnerAndTypes(eq(owner.getId()), any(PoolType.class))).thenReturn(pools);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
         verify(poolRulesMock).createAndEnrichPools(argPool.capture(), any(List.class));
         assertPoolsAreEqual(TestUtil.copyFromSub(s), argPool.getValue());
     }
@@ -1270,7 +1272,7 @@ public class PoolManagerTest {
         // Any positive value is acceptable here
         when(entitlementCurator.getInBlockSize()).thenReturn(50);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
 
         verify(mockPoolCurator).batchDelete(eq(pools), anyCollectionOf(String.class));
         verify(entitlementCurator).batchDeleteByIds(eq(new HashSet<>(Arrays.asList(ent.getId()))));
@@ -1466,7 +1468,7 @@ public class PoolManagerTest {
 
         when(mockPoolCurator.listByOwnerAndTypes(eq(owner.getId()), any(PoolType.class))).thenReturn(pools);
 
-        this.manager.getRefresher(mockSubAdapter).add(owner).run();
+        this.manager.getRefresher(mockSubAdapter, mockProdAdapter).add(owner).run();
 
         // The pool left over from the pre-migrated subscription should be deleted
         // and granted entitlements should be revoked
