+++ server/src/test/java/org/candlepin/resource/GuestIdResourceTest.java
@@ -303,7 +303,7 @@ public class GuestIdResourceTest {
      */
     private class ConsumerResourceForTesting extends ConsumerResource {
         public ConsumerResourceForTesting() {
-            super(null, null, null, null, null, null, null, null, null, null, null, null, null,
+            super(null, null, null, null, null, null, null, null, null, null, null, null, null, null,
                 null, null, null, null, null, null, null, null, null, null, null, null, null,
                 null, null, null, null, consumerEnricher, null, modelTranslator, jobManager);
         }
